requirements.txt
openai==1.35.12
python-dotenv==1.0.1
twitchio==2.10.0